Move plugin helpers
